Remove unused profile setting
This changes removes the configuration option for specifying
a profile to be used. It seems that even by passing profile
in the subscribe hts message, tvheadend always interprets
the value as "htsp". Thus there is no point in keeping it
in the configuration settings.

Some reference for another pvr hts client:
https://github.com/kodi-pvr/pvr.hts/issues/326#issuecomment-325780839